requirements.txt
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
asttokens==3.0.0
asyncer==0.0.8
autogen==0.7.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
comm==0.2.2
debugpy==1.8.12
decorator==5.1.1
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
dotenv==0.9.9
executing==2.2.0
fast-depends==2.4.12
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.68.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
jiter==0.8.2
jupyter_client==8.6.3
jupyter_core==5.7.2
Markdown==3.7
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.2
openai==1.61.1
packaging==24.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.50
proto-plus==1.26.0
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyautogen==0.7.3
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.2.1
regex==2024.11.6
requests==2.32.3
rsa==4.9
-e git+https://github.com/Joelsp19/sar_project@96704ee616ed18fee111f0e89a120a87cd8be0da#egg=sar_project
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
termcolor==2.5.0
tiktoken==0.8.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13
websockets==14.2
